Add static file serving and script for Heroku postbuild
The code adjustments included in this commit allow the express server to serve the frontend static files directly. A wildcard route has been added to serve the main index.html file in the client app. Furthermore, a script for Heroku postbuild has been added to package.json so that the frontend client can also be built automatically when deployed on Heroku.